+++ lib/inspector.dart
@@ -1,0 +1,4 @@
+/// Library for reflection factory inspection.
+library reflection_factory.inspector;
+
+export 'src/reflection_factory_inspector.dart';