+++ test/src/user_reflection_bridge.reflection.g.dart
@@ -1,7 +1,8 @@
 //
 // GENERATED CODE - DO NOT MODIFY BY HAND!
-// BUILDER: "reflection_factory"
+// BUILDER: reflection_factory/1.0.9
 // BUILD COMMAND: dart run build_runner build
+// BUILD TIME: 2021-09-26 15:40:20.005871
 //
 
 part of 'user_reflection_bridge.dart';